Refactor correct_pad_downsample as a util
